Don't store the songs' physical (original) metadata in the database. Instead, load it from the physical files in question, on demand; and store only the necessary data to perform song deduplication in the database. This saves a lot of memory.
Database schema version has been bumped to 3.

As a side effect of these changes, it was discovered that song deduplication never worked for files that were already in the database at startup time; if you added a new media directory that contained files duplicated from existing media directories, they would not be detected as duplicates. Storing the `SimilarityRec`s in the database fixes this.