Some actual handling of errors in the playback thread.
Previously, nearly any audio API error would cause the playback thread to die forever. That's not great, especially given how easy it is to accidentally select a really cranky audio device, or just one that's not even able to produce output. This commit factors out the innermost part of the playback thread, so that the all-powerful `?` operator can be used to handle these errors; now instead of the thread *panicking*, we can retry in a few seconds, after the user has hopefully fixed the problem! Neato!

Still left to do:
- Make it so that the user can't select an audio *input* to play back on (duh?)
- Display errors in the playback thread in the UI, so that the user doesn't have to guess (duh??)
- Resampling and channel conversion routines, so that the JACK audio system can actually be used (issue #5)